+++ package.json
@@ -2,7 +2,8 @@
   "name": "abbie",
   "version": "1.1.4",
   "description": "A simple, clean, javascript logger",
-  "main": "index.js",
+  "main": "main.js",
+  "types": "./typings/",
   "scripts": {
     "test": "npm run jstest && npm run tstest",
     "jstest": "node ./tests/JSTest.js",
@@ -29,7 +30,6 @@
     "@types/node": "^14.14.25",
     "dotenv": "^8.2.0",
     "eslint": "^7.19.0",
-    "nyc": "^15.1.0",
     "prettier": "^2.2.1",
     "typescript": "^4.1.4"
   },
