+++ typings/lib/getName.d.ts
