+++ typings/lib/slash.d.ts
