+++ typings/lib/time.d.ts
