+++ typings/main.d.ts
